Process livesheet in reverse
SPN uses the same sheet for both the URL scraper and the livesheet status updater. This allows us to automatically get URLs straight from chat to Streamwall.

Due to limitations with the Google Sheets library, new links will always be appended to the end of the sheet.

This commit reverses the order in which the livesheet statuses are updated so that the most recent links are processed first.